Split Aave into different modules for history inquiring
Split into blockchain/graph history inquiring modules
