+++ rotkehlchen/chain/ethereum/aave/blockchain.py
@@ -1,21 +1,15 @@
 import logging
-from typing import TYPE_CHECKING, Any, Dict, List, NamedTuple, Optional, Tuple, Union
-
-from gevent.lock import Semaphore
+from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 from rotkehlchen.accounting.structures import Balance
 from rotkehlchen.assets.asset import Asset, EthereumToken
-from rotkehlchen.chain.ethereum.makerdao.common import RAY
 from rotkehlchen.chain.ethereum.structures import AaveEvent
-from rotkehlchen.chain.ethereum.zerion import GIVEN_DEFI_BALANCES
 from rotkehlchen.constants.ethereum import (
-    AAVE_ETH_RESERVE_ADDRESS,
     AAVE_LENDING_POOL,
     ATOKEN_ABI,
     MAX_BLOCKTIME_CACHE,
     ZERO_ADDRESS,
 )
-from rotkehlchen.db.dbhandler import DBHandler
 from rotkehlchen.fval import FVal
 from rotkehlchen.history.price import query_usd_price_zero_if_error
 from rotkehlchen.inquirer import Inquirer
@@ -24,98 +18,17 @@ from rotkehlchen.serialization.deserialize import (
     deserialize_blocknumber,
     deserialize_int_from_hex_or_int,
 )
-from rotkehlchen.typing import ChecksumEthAddress, Timestamp
+from rotkehlchen.typing import ChecksumEthAddress
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils.interfaces import EthereumModule
 from rotkehlchen.utils.misc import hex_or_bytes_to_address, hex_or_bytes_to_int
 
-log = logging.getLogger(__name__)
+from .common import ATOKENS_LIST, AaveHistory, AaveInquirer, _get_reserve_address_decimals
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.ethereum.manager import EthereumManager
+    from rotkehlchen.db.dbhandler import DBHandler
 
-
-ATOKEN_TO_DEPLOYED_BLOCK = {
-    'aETH': 9241088,
-    'aENJ': 10471941,
-    'aDAI': 9241063,
-    'aUSDC': 9241071,
-    'aSUSD': 9241077,
-    'aTUSD': 9241068,
-    'aUSDT': 9241076,
-    'aBUSD': 9747321,
-    'aBAT': 9241085,
-    'aKNC': 9241097,
-    'aLEND': 9241081,
-    'aLINK': 9241091,
-    'aMANA': 9241110,
-    'aMKR': 9241106,
-    'aREP': 9241100,
-    'aREN': 10472062,
-    'aSNX': 9241118,
-    'aWBTC': 9241225,
-    'aYFI': 10748286,
-    'aZRX': 9241114,
-}
-ATOKENS_LIST = [EthereumToken(x) for x in ATOKEN_TO_DEPLOYED_BLOCK]
-
-
-class AaveLendingBalance(NamedTuple):
-    """A balance for Aave lending.
-
-    Asset not included here since it's the key in the map that leads to this structure
-    """
-    balance: Balance
-    apy: FVal
-
-    def serialize(self) -> Dict[str, Union[str, Dict[str, str]]]:
-        return {
-            'balance': self.balance.serialize(),
-            'apy': self.apy.to_percentage(precision=2),
-        }
-
-
-class AaveBorrowingBalance(NamedTuple):
-    """A balance for Aave borrowing.
-
-    Asset not included here since it's the key in the map that leads to this structure
-    """
-    balance: Balance
-    variable_apr: FVal
-    stable_apr: FVal
-
-    def serialize(self) -> Dict[str, Union[str, Dict[str, str]]]:
-        return {
-            'balance': self.balance.serialize(),
-            'variable_apr': self.variable_apr.to_percentage(precision=2),
-            'stable_apr': self.stable_apr.to_percentage(precision=2),
-        }
-
-
-class AaveBalances(NamedTuple):
-    """The Aave balances per account. Using str for symbol since ETH is not a token"""
-    lending: Dict[str, AaveLendingBalance]
-    borrowing: Dict[str, AaveBorrowingBalance]
-
-
-class AaveHistory(NamedTuple):
-    """All events and total interest accrued for all Atoken of an address
-    """
-    events: List[AaveEvent]
-    total_earned: Dict[EthereumToken, Balance]
-
-
-def _get_reserve_address_decimals(symbol: str) -> Tuple[ChecksumEthAddress, int]:
-    """Get the reserve address and the number of decimals for symbol"""
-    if symbol == 'ETH':
-        reserve_address = AAVE_ETH_RESERVE_ADDRESS
-        decimals = 18
-    else:
-        token = EthereumToken(symbol)
-        reserve_address = token.ethereum_address
-        decimals = token.decimals
-
-    return reserve_address, decimals
+log = logging.getLogger(__name__)
 
 
 def _atoken_to_reserve_asset(atoken: EthereumToken) -> Asset:
@@ -125,121 +38,22 @@ def _atoken_to_reserve_asset(atoken: EthereumToken) -> Asset:
     return Asset(reserve_symbol)
 
 
-class Aave(EthereumModule):
-    """Aave integration module
-
-    https://docs.aave.com/developers/developing-on-aave/the-protocol/
-    """
+class AaveBlockchainInquirer(AaveInquirer):
+    """Reads Aave historical data from the chain by querying logs"""
 
     def __init__(
             self,
             ethereum_manager: 'EthereumManager',
-            database: DBHandler,
+            database: 'DBHandler',
             premium: Optional[Premium],
             msg_aggregator: MessagesAggregator,
     ) -> None:
-        self.ethereum = ethereum_manager
-        self.database = database
-        self.msg_aggregator = msg_aggregator
-        self.premium = premium
-        self.history_lock = Semaphore()
-        self.balances_lock = Semaphore()
-
-    def get_balances(
-            self,
-            given_defi_balances: GIVEN_DEFI_BALANCES,
-    ) -> Dict[ChecksumEthAddress, AaveBalances]:
-        with self.balances_lock:
-            return self._get_balances(given_defi_balances)
-
-    def _get_balances(
-            self,
-            given_defi_balances: GIVEN_DEFI_BALANCES,
-    ) -> Dict[ChecksumEthAddress, AaveBalances]:
-        """Retrieves the aave balances
-
-        Receives the defi balances from zerion as an argument. They can either be directly given
-        as the defi balances mapping or as a callable that will retrieve the
-        balances mapping when executed.
-        """
-        aave_balances = {}
-        reserve_cache: Dict[str, Tuple[Any, ...]] = {}
-
-        if isinstance(given_defi_balances, dict):
-            defi_balances = given_defi_balances
-        else:
-            defi_balances = given_defi_balances()
-
-        for account, balance_entries in defi_balances.items():
-            lending_map = {}
-            borrowing_map = {}
-            for balance_entry in balance_entries:
-                if balance_entry.protocol.name != 'Aave':
-                    continue
-
-                # Depending on whether it's asset or debt we find what the reserve asset is
-                if balance_entry.balance_type == 'Asset':
-                    asset = balance_entry.underlying_balances[0]
-                else:
-                    asset = balance_entry.base_balance
-                reserve_address, _ = _get_reserve_address_decimals(asset.token_symbol)
-
-                reserve_data = reserve_cache.get(reserve_address, None)
-                if reserve_data is None:
-                    reserve_data = self.ethereum.call_contract(
-                        contract_address=AAVE_LENDING_POOL.address,
-                        abi=AAVE_LENDING_POOL.abi,
-                        method_name='getReserveData',
-                        arguments=[reserve_address],
-                    )
-                    reserve_cache[balance_entry.base_balance.token_symbol] = reserve_data
-
-                if balance_entry.balance_type == 'Asset':
-                    lending_map[asset.token_symbol] = AaveLendingBalance(
-                        balance=asset.balance,
-                        apy=FVal(reserve_data[4] / RAY),
-                    )
-                else:  # 'Debt'
-                    borrowing_map[asset.token_symbol] = AaveBorrowingBalance(
-                        balance=asset.balance,
-                        variable_apr=FVal(reserve_data[5] / RAY),
-                        stable_apr=FVal(reserve_data[6] / RAY),
-                    )
-
-            if lending_map == {} and borrowing_map == {}:
-                # no aave balances for the account
-                continue
-
-            aave_balances[account] = AaveBalances(lending=lending_map, borrowing=borrowing_map)
-
-        return aave_balances
-
-    def get_history(
-            self,
-            addresses: List[ChecksumEthAddress],
-            reset_db_data: bool,
-            from_timestamp: Timestamp,  # pylint: disable=unused-argument
-            to_timestamp: Timestamp,  # pylint: disable=unused-argument
-    ) -> Dict[ChecksumEthAddress, AaveHistory]:
-        """Detects aave historical data for the given addresses"""
-        result = {}
-        latest_block = self.ethereum.get_latest_block_number()
-        with self.history_lock:
-            if reset_db_data is True:
-                self.database.delete_aave_data()
-
-            for address in addresses:
-                last_query = self.database.get_used_query_range(f'aave_events_{address}')
-                history_results = self.get_history_for_address(
-                    user_address=address,
-                    to_block=latest_block,
-                    given_from_block=last_query[1] + 1 if last_query is not None else None,
-                )
-                if len(history_results.events) == 0:
-                    continue
-                result[address] = history_results
-
-        return result
+        super().__init__(
+            ethereum_manager=ethereum_manager,
+            database=database,
+            premium=premium,
+            msg_aggregator=msg_aggregator,
+        )
 
     def get_history_for_address(
             self,
@@ -481,13 +295,3 @@ class Aave(EthereumModule):
                 ))
 
         return aave_events
-
-    # -- Methods following the EthereumModule interface -- #
-    def on_startup(self) -> None:
-        pass
-
-    def on_account_addition(self, address: ChecksumEthAddress) -> None:
-        pass
-
-    def on_account_removal(self, address: ChecksumEthAddress) -> None:
-        pass
